Decoupling pt.3
expanded the player ship scriptable object so that it includes all the ship components (boost, inventory, mine tool). This means that each script can reference this object instead of having its own serialized field every time, making customizable ships feasible.

This makes it so that the shipmovement script is a requirement for laser, boost, and inventory to work. but like, that's fine.